Return previous result if advance isn't called.
