+++ src/Bedrock.Framework/Protocols/IProtocolWriter.cs
@@ -1,0 +1,9 @@
+﻿using System.Buffers;
+
+namespace Bedrock.Framework.Protocols
+{
+    public interface IProtocolWriter<TMessage>
+    {
+        void WriteMessage(TMessage message, IBufferWriter<byte> output);
+    }
+}
